Add files via upload
This code is about how we can analyze a word that the user has input and break it down to how many number of consonants, vowels and characters are there in the word. I used the traditional approach by going the characters of the word one by one before analyzing them. I tried to use the string.split function to place the word into an array but it seemed like I didn't manage to do so.